wip(fleet-console): Wave 3 — migrate diagrams.ts → @fleet-console/markdown/mermaid
Move the Mermaid diagram hydrator from the codex-private
runtime/fleet-console/core/client/src/codex/markdown/diagrams.ts
into the shared workspace package at runtime/fleet-console/markdown/mermaid.ts.

- Copy diagrams.ts verbatim to markdown/mermaid.ts (import path updated to
  @fleet-console/markdown/core)
- Delete original diagrams.ts and the now-empty markdown/ directory
- reading-controller.ts: switch installDiagramHydrator import to
  @fleet-console/markdown/mermaid
- codex-security-diagrams.test.ts: update all dynamic import specifiers and
  fs.readFile source paths to the new package location
- Security invariants preserved: securityLevel:"strict", htmlLabels:false,
  no bindFunctions, SVG sanitize, mermaid lazy-load pattern unchanged

QA: typecheck (console + file-explorer) PASS; 392 tests PASS
